Update test and move it into Settings folder; delete old ThemeToggle.tsx file from pre-CMR days; delete old PrimaryNav.tsx, make some associated changes for TypeScript purposes
