+++ packages/manager/src/components/PrimaryNav/index.ts
@@ -1,2 +1,2 @@
-import PrimaryNav from './PrimaryNav';
+import PrimaryNav from './PrimaryNav_CMR';
 export default PrimaryNav;
